moved downloadFile to DocumentService
